move all new eval code under preview dir
