+++ packages/phoenix-evals/src/phoenix/evals/__init__.py
@@ -64,19 +64,10 @@ from .templates import (
 from .utils import NOT_PARSABLE, download_benchmark_dataset, emoji_guard
 
 EXPERIMENTAL_MODULES: Dict[str, Dict[str, Any]] = {
-    "phoenix.evals.llm": {
+    "phoenix.evals.preview": {
         "warning_message": (
-            f"\n\n{emoji_guard('⚠️', '!!')} EXPERIMENTAL: The phoenix.evals.llm module and all "
+            f"\n\n{emoji_guard('⚠️', '!!')} EXPERIMENTAL: The phoenix.evals.preview module and all "
             "its components are experimental and subject to change without notice. This code "
-            "should not be used in production."
-        ),
-        "warn_on_access": True,
-        "warning_on_import": True,
-    },
-    "phoenix.evals.templating": {
-        "warning_message": (
-            f"\n\n{emoji_guard('⚠️', '!!')} EXPERIMENTAL: The phoenix.evals.templating module and "
-            "all its components are experimental and subject to change without notice. This code "
             "should not be used in production."
         ),
         "warn_on_access": True,
@@ -175,16 +166,16 @@ class ExperimentalModuleLoader(Loader):
 
 sys.meta_path.insert(0, ExperimentalModuleFinder())
 
-from . import llm  # noqa: E402
+from . import preview  # noqa: E402
 
 __version__ = version("arize-phoenix-evals")
 
 __all__ = [
     "compute_precisions_at_k",
     "download_benchmark_dataset",
-    "llm",
     "llm_classify",
     "llm_generate",
+    "preview",
     "OpenAIModel",
     "AnthropicModel",
     "GeminiModel",
