+++ packages/phoenix-evals/tests/phoenix/evals/preview/llm/test_wrapper.py
@@ -3,7 +3,7 @@ from typing import Any, Dict, List, Union
 
 import pytest
 
-from phoenix.evals.llm.wrapper import generate_classification_schema
+from phoenix.evals.preview.llm.wrapper import generate_classification_schema
 
 
 class TestGenerateClassificationSchema:
